requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.7.0
google-generativeai
pinecone
python-dotenv
pypdf
mcp>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
